+++ src/app/App.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { StatusBar } from "expo-status-bar";
 import { Text } from "react-native";
 import {
   GoogleAuthProvider,
@@ -11,7 +10,6 @@ import * as Google from "expo-auth-session/providers/google";
 import * as WebBrowser from "expo-web-browser";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { User } from "firebase/auth";
-import { StyleSheet, View } from "react-native";
 
 // Imports for Navigation
 import { NavigationContainer } from "@react-navigation/native";
@@ -29,8 +27,6 @@ import { registerRootComponent } from "expo";
 
 WebBrowser.maybeCompleteAuthSession();
 
-const Stack = createStackNavigator<RootStackParamList>();
-
 // Types for navigation handling
 // Should navigation be handled in a separate file??
 type RootStackParamList = {
@@ -40,15 +36,9 @@ type RootStackParamList = {
   OrderMenu: undefined;
 };
 
+const Stack = createStackNavigator<RootStackParamList>();
+
 function App() {
-  const [request, response, promptAsync] = Google.useAuthRequest({
-    iosClientId:
-      "983400403511-gi5mo0akb89fcecaivk4q509c63hrvtl.apps.googleusercontent.com",
-    androidClientId:
-      "983400403511-i43set67i4o1e3kb7fl91vrh9r6aemcb.apps.googleusercontent.com",
-    redirectUri:
-      "com.googleusercontent.apps.983400403511-gi5mo0akb89fcecaivk4q509c63hrvtl:/oauth2redirect/google",
-  });
 
   const [fontsLoaded] = useFonts({
     "Kaisei-Regular": KaiseiRegular,
@@ -59,11 +49,11 @@ function App() {
 
   const checkLocalUser = async () => {
     try {
-      setLoading(true);
-      const userJSON = await AsyncStorage.getItem("@user");
-      const userData = userJSON != null ? JSON.parse(userJSON) : null;
-      console.log("local storage:", userData);
-      setUserInfo(userData);
+      // setLoading(true);
+      // const userJSON = await AsyncStorage.getItem("@user");
+      // const userData = userJSON != null ? JSON.parse(userJSON) : null;
+      // console.log("local storage:", userData);
+      // setUserInfo(userData);
     } catch (e) {
       alert(e);
     } finally {
@@ -71,13 +61,6 @@ function App() {
     }
   };
 
-  React.useEffect(() => {
-    if (response?.type === "success") {
-      const { id_token } = response.params;
-      const credential = GoogleAuthProvider.credential(id_token);
-      signInWithCredential(auth, credential);
-    }
-  }, [response]);
 
   React.useEffect(() => {
     checkLocalUser();
@@ -104,10 +87,10 @@ function App() {
     <NavigationContainer>
       <Stack.Navigator initialRouteName={userInfo ? "OrderMenu" : "Login"}>
         <Stack.Screen name="Login" options={{ title: "Login to Wild Knight" }}>
-          {(props) => <Login {...props} promptAsync={promptAsync} />}
+          {(props) => <Login {...props} />}
         </Stack.Screen>
         <Stack.Screen name="SignUp">
-          {(props) => <SignUp {...props} promptAsync={promptAsync} />}
+          {(props) => <SignUp {...props} />}
         </Stack.Screen>
         <Stack.Screen name="ForgotPassword" component={ForgotPassword} />
         <Stack.Screen name="OrderMenu">{(props) => <OrderMenu />}</Stack.Screen>
